tein moduuli 3

+++ moduuli4/Tehtävä2.py
@@ -2,7 +2,7 @@ tuuma = 2.54
 
 
 while True:
-    T = int(input("anna tuumat: "))
+    T = float(input("anna tuumat: "))
     tulos = T * tuuma
     if tulos <= 0:
         break
